backend/requirements-common.txt
flask==2.0.1
pyjwt==2.1.0
passlib==1.7.4
bcrypt==3.2.0
requests==2.25.1
